1. Creation of a Settings component ans used it for landing page and options page. 2. Styles fixes.
