+++ src/components/TimerCards/HabbitReminder.tsx
@@ -60,6 +60,7 @@ const styles = StyleSheet.create({
     title: {
         fontSize: 17,
         alignSelf: 'center',
+        textAlign: 'center',
         marginBottom: 10
     },
     habbit: {
