+++ src/navigation/index.tsx
@@ -15,7 +15,6 @@ import { switchTheme } from '../redux/actions'
 
 // == COMPONENT
 import LandingPage from '../screens/LandingPage';
-import HomePage from '../screens/HomePage';
 import WorkCard from '../components/TimerCards/WorkCard';
 import PauseCard from '../components/TimerCards/PauseCard'
 import Options from '../screens/Options';
@@ -48,47 +47,15 @@ export const AppNavigation: React.FC<StackParamsList> = () => {
                             backgroundColor: Dark.gold,
                         },
                         headerRight: () => (
-                            <>
-                                <Switch 
-                                    onValueChange={(() => setIsToggle(!isToggle))}
-                                    value={isToggle}
-                                    thumbColor={isToggle ? Dark.dark : Light.primary}
-                                    trackColor={{false: Light.secondary, true: Dark.primary }}
-                                />
-                                <Pressable
-                                    onPress={() => navigation.navigate('Settings')}
-                                >
-                                    <MaterialIcons name="settings" size={24} color={Dark.dark} />
-                                </Pressable>
-                            </>
+                            <Switch 
+                                onValueChange={(() => setIsToggle(!isToggle))}
+                                value={isToggle}
+                                thumbColor={isToggle ? Dark.dark : Light.primary}
+                                trackColor={{false: Light.secondary, true: Dark.primary }}
+                            />
                         )
                     })}
                 />
-                {/* <Stack.Screen
-                    name="HomePage"
-                    component={HomePage}
-                    options={({ navigation }) => ({
-                        title: 'Accueil',
-                        headerStyle: {
-                            backgroundColor: Dark.gold
-                        },
-                        headerRight: () => (
-                            <>
-                                <Switch 
-                                    onValueChange={() => setIsToggle(!isToggle)}
-                                    value={isToggle}
-                                    thumbColor={isToggle ? Dark.dark : Light.primary}
-                                    trackColor={{false: Light.secondary, true: Dark.primary }}
-                                />
-                                <Pressable
-                                    onPress={() => navigation.navigate('Settings')}
-                                >
-                                    <MaterialIcons name="settings" size={24} color={Dark.dark} />
-                                </Pressable>
-                            </>
-                        )
-                    })}
-                /> */}
                 <Stack.Screen
                     name="WorkCard"
                     component={WorkCard}
@@ -139,11 +106,24 @@ export const AppNavigation: React.FC<StackParamsList> = () => {
                         )
                     })}
                 />
-                {/* <Stack.Screen
+                <Stack.Screen
                     name="Settings"
                     component={Options}
-                    
-                /> */}
+                    options={({ navigation }) => ({
+                        title: 'Options',
+                        headerStyle: {
+                            backgroundColor: Dark.gold
+                        },
+                        headerRight: () => (
+                            <Switch 
+                                onValueChange={() => setIsToggle(!isToggle)}
+                                value={isToggle}
+                                thumbColor={isToggle ? Dark.dark : Light.primary}
+                                trackColor={{false: Light.secondary, true: Dark.primary }}
+                            />
+                        )
+                    })}
+                />
             </Stack.Navigator>
         </NavigationContainer>
     )
